Refactor code and change folder name
