+++ README.md
@@ -13,6 +13,7 @@ A set of examples to help demontrate common configurations using Mocha. They sho
 * [Third party reporter](third-party-reporter/)
 * [TypeScript application](typescript/)
 * [Vue and Puppeteer application](vue-puppeteer/)
+* [Express REST API](express-rest-api/)
 
 ## Requirements
 
@@ -24,7 +25,6 @@ Feel free to send a branch to supply an MCVE from one of the applications.
 
 ## Potential setups to add in future
 
-- Rest API
 - Angular
 - async setups (i.e. `--delay`)
 - Pipe json output to file
