+++ express-rest-api/README.md
@@ -1,9 +1,7 @@
-# REST API example with Restify
+# Express REST API example
 
 This example demonstrate how you can create test code using Mocha.js
 [Express](https://expressjs.com) used to create a server and [supertest](https://github.com/visionmedia/supertest) used for HTTP assertions, then Mocha validates the result. 
 
 ## Commands
-- `npm install` - install npm packages
-- `npm build` - build the mocha test code 
 - `npm test` - run the mocha test code
