+++ express-rest-api/server/app.js
@@ -15,19 +15,13 @@ app.use('/users', usersRouter);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
-  next(createError(404));
+  res.status(404).send("Sorry can't find that!")
 });
 
-// error handler
-app.use(function(err, req, res, next) {
-  // set locals, only providing error in development
-  res.locals.message = err.message;
-  res.locals.error = req.app.get('env') === 'development' ? err : {};
-
-  console.log(err)
-  // render the error page
-  res.status(err.status || 500);
-  res.send('error');
-});
+// catch 500 and forward to error handler
+app.use(function (err, req, res, next) {
+  console.error(err.stack)
+  res.status(500).send('Something broke!')
+})
 
 module.exports = app;
