+++ express-rest-api/server/routes/index.js
@@ -3,7 +3,6 @@ var router = express.Router();
 
 /* GET home page. */
 router.get('/', function(req, res, next) {
-  //res.render('index', { title: 'Express' });
   res.send('<p>some html</p>');
 });
